[Pal] Refactor simple types, part 1 (bools, enums, strings)
There's no reason for PAL API to have custom types for these, which
effectively were erasing types from PAL API. This commit replaces them
with proper C types.

The rest of the types will be changed separately.

In order to make the PAL API more regular, creating eventfds is now done
with two calls on the host: eventfd() with the initial value of 0, and
write() with the desired initial value.

Also, DkSegmentRegister{Get,Set} were renamed to DkSegmentBase{Get,Set}.

Signed-off-by: Michał Kowalczyk <mkow@invisiblethingslab.com>